// src/proto/dapr/proto/runtime/v1/dapr.ts
export interface TryLockRequest {
  /** Required. The lock store name,e.g. `redis`. */
  storeName: string;
  /**
   * Required. resource_id is the lock key. e.g. `order_id_111`
   * It stands for "which resource I want to protect"
   */
  resourceId: string;
  /**
   * Required. lock_owner indicate the identifier of lock owner.
   * You can generate a uuid as lock_owner.For example,in golang:
   *
   * req.LockOwner = uuid.New().String()
   *
   * This field is per request,not per process,so it is different for each request,
   * which aims to prevent multi-thread in the same process trying the same lock concurrently.
   *
   * The reason why we don't make it automatically generated is:
   * 1. If it is automatically generated,there must be a 'my_lock_owner_id' field in the response.
   * This name is so weird that we think it is inappropriate to put it into the api spec
   * 2. If we change the field 'my_lock_owner_id' in the response to 'lock_owner',which means the current lock owner of this lock,
   * we find that in some lock services users can't get the current lock owner.Actually users don't need it at all.
   * 3. When reentrant lock is needed,the existing lock_owner is required to identify client and check "whether this client can reenter this lock".
   * So this field in the request shouldn't be removed.
   */
  lockOwner: string;
  /** Required. The time before expiry.The time unit is second. */
  expiryInSeconds: number;
}

export interface TryLockResponse {
  success: boolean;
}

export interface UnlockRequest {
  storeName: string;
  /** resource_id is the lock key. */
  resourceId: string;
  lockOwner: string;
}
// ...
export const TryLockRequest = {
  encode(message: TryLockRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.storeName !== "") {
      writer.uint32(10).string(message.storeName);
    }
    if (message.resourceId !== "") {
      writer.uint32(18).string(message.resourceId);
    }
    if (message.lockOwner !== "") {
      writer.uint32(26).string(message.lockOwner);
    }
    if (message.expiryInSeconds !== 0) {
      writer.uint32(32).int32(message.expiryInSeconds);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TryLockRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTryLockRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.storeName = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.resourceId = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.lockOwner = reader.string();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.expiryInSeconds = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<TryLockRequest, Uint8Array>
  async *encodeTransform(
    source: AsyncIterable<TryLockRequest | TryLockRequest[]> | Iterable<TryLockRequest | TryLockRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [TryLockRequest.encode(p).finish()];
        }
      } else {
        yield* [TryLockRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, TryLockRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<TryLockRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [TryLockRequest.decode(p)];
        }
      } else {
        yield* [TryLockRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): TryLockRequest {
    return {
      storeName: isSet(object.storeName) ? globalThis.String(object.storeName) : "",
      resourceId: isSet(object.resourceId) ? globalThis.String(object.resourceId) : "",
      lockOwner: isSet(object.lockOwner) ? globalThis.String(object.lockOwner) : "",
      expiryInSeconds: isSet(object.expiryInSeconds) ? globalThis.Number(object.expiryInSeconds) : 0,
    };
  },

  toJSON(message: TryLockRequest): unknown {
    const obj: any = {};
    if (message.storeName !== "") {
      obj.storeName = message.storeName;
    }
    if (message.resourceId !== "") {
      obj.resourceId = message.resourceId;
    }
    if (message.lockOwner !== "") {
      obj.lockOwner = message.lockOwner;
    }
    if (message.expiryInSeconds !== 0) {
      obj.expiryInSeconds = Math.round(message.expiryInSeconds);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TryLockRequest>, I>>(base?: I): TryLockRequest {
    return TryLockRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TryLockRequest>, I>>(object: I): TryLockRequest {
    const message = createBaseTryLockRequest();
    message.storeName = object.storeName ?? "";
    message.resourceId = object.resourceId ?? "";
    message.lockOwner = object.lockOwner ?? "";
    message.expiryInSeconds = object.expiryInSeconds ?? 0;
    return message;
  },
};

function createBaseTryLockResponse(): TryLockResponse {
  return { success: false };
}

export const TryLockResponse = {
  encode(message: TryLockResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.success !== false) {
      writer.uint32(8).bool(message.success);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TryLockResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTryLockResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.success = reader.bool();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<TryLockResponse, Uint8Array>
  async *encodeTransform(
    source: AsyncIterable<TryLockResponse | TryLockResponse[]> | Iterable<TryLockResponse | TryLockResponse[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [TryLockResponse.encode(p).finish()];
        }
      } else {
        yield* [TryLockResponse.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, TryLockResponse>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<TryLockResponse> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [TryLockResponse.decode(p)];
        }
      } else {
        yield* [TryLockResponse.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): TryLockResponse {
    return { success: isSet(object.success) ? globalThis.Boolean(object.success) : false };
  },

  toJSON(message: TryLockResponse): unknown {
    const obj: any = {};
    if (message.success !== false) {
      obj.success = message.success;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TryLockResponse>, I>>(base?: I): TryLockResponse {
    return TryLockResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TryLockResponse>, I>>(object: I): TryLockResponse {
    const message = createBaseTryLockResponse();
    message.success = object.success ?? false;
    return message;
  },
};

function createBaseUnlockRequest(): UnlockRequest {
  return { storeName: "", resourceId: "", lockOwner: "" };
}

export const UnlockRequest = {
  encode(message: UnlockRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.storeName !== "") {
      writer.uint32(10).string(message.storeName);
    }
    if (message.resourceId !== "") {
      writer.uint32(18).string(message.resourceId);
    }
    if (message.lockOwner !== "") {
      writer.uint32(26).string(message.lockOwner);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): UnlockRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnlockRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.storeName = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.resourceId = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.lockOwner = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<UnlockRequest, Uint8Array>
  async *encodeTransform(
    source: AsyncIterable<UnlockRequest | UnlockRequest[]> | Iterable<UnlockRequest | UnlockRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [UnlockRequest.encode(p).finish()];
        }
      } else {
        yield* [UnlockRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, UnlockRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<UnlockRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [UnlockRequest.decode(p)];
        }
      } else {
        yield* [UnlockRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): UnlockRequest {
    return {
      storeName: isSet(object.storeName) ? globalThis.String(object.storeName) : "",
      resourceId: isSet(object.resourceId) ? globalThis.String(object.resourceId) : "",
      lockOwner: isSet(object.lockOwner) ? globalThis.String(object.lockOwner) : "",
    };
  },

  toJSON(message: UnlockRequest): unknown {
    const obj: any = {};
    if (message.storeName !== "") {
      obj.storeName = message.storeName;
    }
    if (message.resourceId !== "") {
      obj.resourceId = message.resourceId;
    }
    if (message.lockOwner !== "") {
      obj.lockOwner = message.lockOwner;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UnlockRequest>, I>>(base?: I): UnlockRequest {
    return UnlockRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UnlockRequest>, I>>(object: I): UnlockRequest {
    const message = createBaseUnlockRequest();
    message.storeName = object.storeName ?? "";
    message.resourceId = object.resourceId ?? "";
    message.lockOwner = object.lockOwner ?? "";
    return message;
  },
};